Support Unicode emoji characters in BitmapFont
The .NET Char type currently used to identify a glyph in a BitmapFont is
limited to 16 bits, while some Unicode characters (surrogate pairs,
which include Emoji) require 32 bits.

The fix is to use the Int32 type to represent a Unicode character, and
to parse UTF-16 .NET input strings as 32-bit Unicode code point
sequences.