Added Rough Group Layer Support
Added the ability to load files with groups. Rendering system still doesn't understand layer groups yet, so this allows groups to be initially loaded, but the sub layers of the group are extracted to the layer root, things like offsets for the group are applied directly to the sub layers, and the groups themselves are removed.