Add defensive check for IsDiposed when getting texture
Added additional condition to check if the `_whitePixelTexture` has been disposed of before reusing.  This prevents `ObjectDisposedExceptions` for some platforms like Android, which can occur if the app is in the background for too long, causing the android activity to stop but the application to still be running.  If this occurs, when switching back to the application, the VRAM may be cleared and the texture disposed.