Mark PitchUp and PitchDown as obsolete
Pitch was introduced in PR #907. While the intentions of the implementation were good, moving forward it is going to start causing technical debt to keep it as a property for all camera types. Having pitch doesn't make sense for an Orthographic camera, and would be better served as an implementation into an isometric or even a perspective camera implementation in the future.

This is marked as a compiler warning (not an error), so existing code will continue to work. The property will be removed in the next major version.