Revert back to using indexer to add kerning pairs
Sometimes BMFont generation can create duplicate kerning pairs.  When this happens, the file is technically malformed, however previous behavior was to use the indexer of the dictionary to add the pairs, which just overwrite them.

Reverted back to using indexer and update the processor to detect duplicates and warn users now during build